Demos hooked up with nice .astro styling
