+++ packages/emoji-blast-site/src/components/Demo.module.css
@@ -1,0 +1,3 @@
+.demo {
+	padding-bottom: 2rem;
+}
