+++ packages/emoji-blast-site/src/data/slugify.ts
@@ -1,0 +1,3 @@
+export function slugify(text: string) {
+	return text.toLowerCase().replace(" ", "-");
+}
